+++ index.test.js
@@ -149,21 +149,17 @@ describe('Equivalency', () => {
     });
   });
 
-  describe('equivalent (common en diacritics)', () => {
-    it('should return true when inputs differ solely by common en diacritics', () => {
-      const instance = new Equivalency().doesntMatter(
-        Equivalency.en.COMMON_DIACRITICS
-      );
+  describe('equivalent (common diacritics)', () => {
+    it('should return true when inputs differ solely by common diacritics', () => {
+      const instance = new Equivalency().doesntMatter(Equivalency.DIACRITICS);
       const inputs = [['â', 'a']];
       inputs.forEach(([s1, s2]) => {
         expect(instance.equivalent(s1, s2)).toEqual({ isEquivalent: true });
       });
     });
 
-    it('should return false when inputs differ other than by common en diacritics', () => {
-      const instance = new Equivalency().doesntMatter(
-        Equivalency.en.COMMON_DIACRITICS
-      );
+    it('should return false when inputs differ other than by common diacritics', () => {
+      const instance = new Equivalency().doesntMatter(Equivalency.DIACRITICS);
       const inputs = [['âb', 'âc']];
       inputs.forEach(([s1, s2]) => {
         expect(instance.equivalent(s1, s2)).toEqual({ isEquivalent: false });
@@ -177,7 +173,7 @@ describe('Equivalency', () => {
         .doesntMatter(Equivalency.CAPITALIZATION)
         .doesntMatter(Equivalency.en.COMMON_PUNCTUATION)
         .doesntMatter(Equivalency.en.COMMON_SYMBOLS)
-        .doesntMatter(Equivalency.en.COMMON_DIACRITICS);
+        .doesntMatter(Equivalency.DIACRITICS);
 
       const { isEquivalent } = enEquivalency.equivalent(
         'àâäçèéêíïîñóöüÀÂÄÇÈÉÊÍÏÎÑÓÖÜ',
@@ -286,7 +282,7 @@ describe('Real-world usage', () => {
       .doesntMatter(Equivalency.CAPITALIZATION)
       .doesntMatter(Equivalency.en.COMMON_PUNCTUATION)
       .doesntMatter(Equivalency.en.COMMON_SYMBOLS)
-      .doesntMatter(Equivalency.ACCENTS)
+      .doesntMatter(Equivalency.DIACRITICS)
       .matters('-');
 
     it('should mark candidates equivalent that we want to count as equivalent', () => {
