+++ lib/en.js
@@ -1,4 +1,4 @@
-const { replaceAll, FunctionRule } = require('./definitions');
+const { FunctionRule } = require('./definitions');
 const normalize = require('./normalize');
 
 // These two lists contain all ASCII symbols, where symbol is defined as a
@@ -9,28 +9,6 @@ const ASCII_SYMBOLS = `$#&%()*+/\\<=>@[]^_{|}~`;
 const COMMON_NONASCII_PUNCTUATION = `…“”‘’`;
 const COMMON_PUNCTUATION = ASCII_PUNCTUATION + COMMON_NONASCII_PUNCTUATION;
 const COMMON_SYMBOLS = ASCII_SYMBOLS;
-
-const commonCombiningDiacritics = [
-  '\u0300', // grave accent
-  '\u0301', // acute accent
-  '\u0302', // circumflex
-  '\u0303', // tilde
-  '\u0308', // diaresis
-  '\u0327', // cedilla
-];
-
-const COMMON_DIACRITICS = new FunctionRule((s1, s2) => {
-  let s1prime = normalize(s1, 'NFD');
-  let s2prime = normalize(s2, 'NFD');
-
-  commonCombiningDiacritics.forEach(c => {
-    s1prime = replaceAll(s1prime, c);
-    s2prime = replaceAll(s2prime, c);
-  });
-
-  // FIXME: should be restored to whatever normalization they came in with.
-  return [normalize(s1prime, 'NFC'), normalize(s2prime, 'NFC')];
-});
 
 const INFINITIVE_VERBS = new FunctionRule((s1, s2) => {
   let s1prime = normalize(s1, 'NFD');
@@ -49,6 +27,5 @@ module.exports = {
   ASCII_SYMBOLS,
   COMMON_PUNCTUATION,
   COMMON_SYMBOLS,
-  COMMON_DIACRITICS,
   INFINITIVE_VERBS,
 };
