+++ lib/index.js
@@ -26,15 +26,20 @@ const whitespaceDifferences = (s1, s2) => {
 };
 const whitespaceDifferencesRule = new FunctionRule(whitespaceDifferences);
 
-// TODO: Could generalize to diacritics, i.e., Unicode category "nonspacing
-// mark" (see http://www.fileformat.info/info/unicode/category/Mn/list.htm)
-const combiningAccents = ['\u0300', '\u0301', '\u0302'];
+const commonCombiningDiacritics = [
+  '\u0300', // grave accent
+  '\u0301', // acute accent
+  '\u0302', // circumflex
+  '\u0303', // tilde
+  '\u0308', // diaresis
+  '\u0327', // cedilla
+];
 
-const accentsRule = new FunctionRule((s1, s2) => {
+const diacriticsRule = new FunctionRule((s1, s2) => {
   let s1prime = normalize(s1, 'NFD');
   let s2prime = normalize(s2, 'NFD');
 
-  combiningAccents.forEach(c => {
+  commonCombiningDiacritics.forEach(c => {
     s1prime = replaceAll(s1prime, c);
     s2prime = replaceAll(s2prime, c);
   });
@@ -70,6 +75,6 @@ module.exports = {
   CAPITALIZATION: capitalizationRule,
   UNICODE_NORMALIZATION: unicodeNormalizationRule,
   WHITESPACE_DIFFERENCES: whitespaceDifferencesRule,
-  ACCENTS: accentsRule,
+  DIACRITICS: diacriticsRule,
   wordPrefix: wordPrefixRuleCreator,
 };
